% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_package.R
\docType{package}
\name{stantargets-package}
\alias{stantargets-package}
\title{targets: Targets Archetypes for Stan}
\description{
The \code{targets} package is a pipeline toolkit that brings together
function-oriented programming and Make-like declarative workflows for
Statistics and data science in R. The \code{stantargets} package provides
convenient user-side functions to create Bayesian data analysis
workflows with \code{targets} and Stan. These functions reduce
cognitive overhead and micromanagement so the user can focus
more on model development than software development.
\code{stantargets} invokes Stan through the \code{cmdstanr} package:
\url{https://mc-stan.org/cmdstanr}.
}
