% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_compile.R
\name{tar_stan_compile_run}
\alias{tar_stan_compile_run}
\title{Compile a Stan model and return the model file.}
\usage{
tar_stan_compile_run(
  stan_file,
  quiet = TRUE,
  dir = NULL,
  include_paths = NULL,
  cpp_options = list(),
  stanc_options = list(),
  force_recompile = FALSE
)
}
\arguments{
\item{stan_file}{The path to a \code{.stan} file containing a Stan program. The
helper function \code{\link[cmdstanr:write_stan_file]{write_stan_file()}} is provided for cases when it is
more convenient to specify the Stan program as a string.}

\item{quiet}{\code{quiet} argument to \verb{$compile()}.}

\item{dir}{\code{dir} argument to \verb{$compile()}.}

\item{include_paths}{\code{include_paths} argument to \verb{$compile()}.}

\item{cpp_options}{\code{cpp_options} argument to \verb{$compile()}.}

\item{stanc_options}{\code{stanc_options} argument to \verb{$compile()}.}

\item{force_recompile}{\code{force_recompile} argument to \verb{$compile()}.}
}
\value{
Character of length 1, the value of \code{stan_file}
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
