% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_example_data.R
\name{tar_stan_example_data}
\alias{tar_stan_example_data}
\title{Example data for \code{\link[=tar_stan_example_file]{tar_stan_example_file()}}}
\usage{
tar_stan_example_data(n = 10L)
}
\arguments{
\item{n}{Integer of length 1, number of data points.}
}
\value{
List, dataset compatible with the model file from
\code{\link[=tar_stan_example_file]{tar_stan_example_file()}}.
}
\description{
An example dataset compatible with the model file
from \code{\link[=tar_stan_example_file]{tar_stan_example_file()}}.
}
\examples{
tar_stan_example_data()
}
