% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_example_file.R
\name{tar_stan_example_file}
\alias{tar_stan_example_file}
\title{Write an example Stan model file.}
\usage{
tar_stan_example_file(path = tempfile(pattern = "", fileext = ".stan"))
}
\arguments{
\item{path}{Character of length 1, file path to write the model file.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Overwrites the file at \code{path} with a built-in example
Stan model file.
}
\examples{
path <- tempfile(pattern = "", fileext = ".stan")
tar_stan_example_file(path = path)
writeLines(readLines(path))
}
\seealso{
Other examples: 
\code{\link{tar_stan_example_data}()}
}
\concept{examples}
