% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_gq_rep_summary.R
\name{tar_stan_gq_rep_summary_run}
\alias{tar_stan_gq_rep_summary_run}
\title{Run a Stan model and return only the summaries.}
\usage{
tar_stan_gq_rep_summary_run(
  stan_file,
  stan_name,
  stan_path,
  data,
  fitted_params,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  output_dir,
  sig_figs,
  parallel_chains,
  threads_per_chain,
  copy_data,
  variables,
  summaries,
  summary_args
)
}
\arguments{
\item{data}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{fitted_params}{Symbol, name of a \code{CmdStanFit} object
computed in a previous target: for example, the
\verb{*_mcmc_*} target from \code{\link[=tar_stan_mcmc]{tar_stan_mcmc()}}. Must be a subclass
that \verb{$generate_quantities()} can accept as \code{fitted_params}.}

\item{compile}{Do compilation? The default is \code{TRUE}. If \code{FALSE}
compilation can be done later via the \code{\link[cmdstanr:model-method-compile]{$compile()}}
method.}

\item{quiet}{(logical) Should the verbose output from CmdStan during
compilation be suppressed? The default is \code{TRUE}, but if you encounter an
error we recommend trying again with \code{quiet=FALSE} to see more of the
output.}

\item{dir}{(string) The path to the directory in which to store the CmdStan
executable (or \code{.hpp} file if using \verb{$save_hpp_file()}). The default is the
same location as the Stan program.}

\item{include_paths}{(character vector) Paths to directories where Stan
should look for files specified in \verb{#include} directives in the Stan
program.}

\item{cpp_options}{(list) Any makefile options to be used when compiling the
model (\code{STAN_THREADS}, \code{STAN_MPI}, \code{STAN_OPENCL}, etc.). Anything you would
otherwise write in the \code{make/local} file.}

\item{stanc_options}{(list) Any Stan-to-C++ transpiler options to be used
when compiling the model. See the \strong{Examples} section below as well as the
\code{stanc} chapter of the CmdStan Guide for more details on available options:
https://mc-stan.org/docs/cmdstan-guide/stanc.html.}

\item{force_recompile}{(logical) Should the model be recompiled even if was
not modified since last compiled. The default is \code{FALSE}.}

\item{seed}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{output_dir}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{sig_figs}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{parallel_chains}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{threads_per_chain}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{variables}{(character vector) The variables to include.}

\item{summaries}{Optional list of summary functions passed to \code{...} in
\code{posterior::summarize_draws()} through \verb{$summary()}
on the \code{CmdStanFit} object.}

\item{summary_args}{Optional list of summary function arguments passed to
\code{.args} in \code{posterior::summarize_draws()} through \verb{$summary()}
on the \code{CmdStanFit} object.}
}
\value{
A data frame of posterior summaries.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
