% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_gq.R
\name{tar_stan_gq_run}
\alias{tar_stan_gq_run}
\title{Compile and run a Stan model and return the \code{CmdStanFit} object.}
\usage{
tar_stan_gq_run(
  stan_file,
  data,
  fitted_params,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  output_dir,
  sig_figs,
  parallel_chains,
  threads_per_chain,
  variables
)
}
\arguments{
\item{stan_file}{The path to a \code{.stan} file containing a Stan program. The
helper function \code{\link[cmdstanr:write_stan_file]{write_stan_file()}} is provided for cases when it is
more convenient to specify the Stan program as a string.}

\item{data}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{fitted_params}{(multiple options) The parameter draws to use. One of
the following:
\itemize{
\item A \link[cmdstanr]{CmdStanMCMC} or \link[cmdstanr]{CmdStanVB} fitted model object.
\item A \link[posterior:draws_array]{posterior::draws_array} (for MCMC) or \link[posterior:draws_matrix]{posterior::draws_matrix} (for
VB) object returned by CmdStanR's \code{\link[cmdstanr:fit-method-draws]{$draws()}} method.
\item A character vector of paths to CmdStan CSV output files.
}}

\item{compile}{Do compilation? The default is \code{TRUE}. If \code{FALSE}
compilation can be done later via the \code{\link[cmdstanr:model-method-compile]{$compile()}}
method.}

\item{seed}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{output_dir}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{sig_figs}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{parallel_chains}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}

\item{threads_per_chain}{Same
as for the \code{\link[cmdstanr:model-method-sample]{$sample()}} method.}
}
\value{
A \code{CmdStanFit} object.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
