% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_mcmc_gq.R
\name{tar_stan_mcmc_gq_run}
\alias{tar_stan_mcmc_gq_run}
\title{Compile and run a Stan model and return the \code{CmdStanFit} object.}
\usage{
tar_stan_mcmc_gq_run(
  stan_file,
  data,
  fitted_params,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  output_dir,
  sig_figs,
  parallel_chains,
  threads_per_chain,
  variables
)
}
\arguments{
\item{stan_file}{Character, Stan model file.}

\item{data}{List of data to pass to the Stan model.}

\item{fitted_params}{An existing \code{CmdStanMCMC} with all draws loaded
into memory.}

\item{compile}{Character of length 1. If \code{"original"}, then
\code{cmdstan} will compile the source file right before running
it (or skip compilation if the binary is up to date). This
assumes the worker has access to the file. If the worker
is running on a remote computer that does not have access
to the model file, set to \code{"copy"} instead. \code{compile = "copy"}
means the pipeline will read the lines of the original Stan model file
and send them to the worker. The worker writes the lines
to a local copy and compiles the model from there, so it
no longer needs access to the original Stan model file on your
local machine. However, as a result, the Stan model re-compiles
every time the main target reruns.}

\item{seed}{\code{seed} argument to \verb{$generate_quantities()}.}

\item{output_dir}{\code{output_dir} argument to \verb{$generate_quantities()}.}

\item{sig_figs}{\code{sig_figs} argument to \verb{$generate_quantities()}.}

\item{parallel_chains}{\code{parallel_chains} argument to
\verb{$generate_quantities()}.}

\item{threads_per_chain}{\code{threads_per_chain} argument to
\verb{$generate_quantities()}.}

\item{variables}{\code{variables} argument to \verb{$draws()} and \verb{$summary()}
on the \code{CmdStanGQ} object.}
}
\value{
A \code{CmdStanFit} object.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
