% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{tar_stan_output}
\alias{tar_stan_output}
\title{Post-process Stan output}
\usage{
tar_stan_output(
  fit,
  output,
  summaries,
  summary_args,
  variables,
  inc_warmup,
  data,
  data_copy
)
}
\arguments{
\item{fit}{A Stan fit object.}

\item{output}{Type of output to create, either \code{"summaries"},
\code{"draws"}, or \code{"diagnostics"}.}

\item{summaries}{Optional list of summary functions passed to \code{...} in
\code{posterior::summarize_draws()} through \verb{$summary()}
on the \code{CmdStanFit} object.}

\item{summary_args}{Optional list of summary function arguments passed to
\code{.args} in \code{posterior::summarize_draws()} through \verb{$summary()}
on the \code{CmdStanFit} object.}

\item{variables}{(character vector) The variables to include.}

\item{inc_warmup}{Logical, whether to include the warmup draws.}

\item{data}{List, Stan dataset.}

\item{data_copy}{Character vector of names of scalars in \code{data}.
These values will be inserted as columns in the output data frame
for each rep. Useful for simulation studies where you want to
check the results against some "true value" in the data. See the
\code{mcmc_rep} vignette for an example.}
}
\value{
A data frame of user-friendly Stan output.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
