% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{tar_stan_rep_data_batch}
\alias{tar_stan_rep_data_batch}
\title{Generate a batch of data}
\usage{
tar_stan_rep_data_batch(reps, command)
}
\arguments{
\item{reps}{Positive integer of length 1, number of reps to run.}

\item{command}{R code to run to generate one dataset.}
}
\value{
A list of Stan datasets containing data and dataset IDs.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
tar_stan_rep_data_batch(2, tar_stan_example_data())
}
\keyword{internal}
