% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_summary.R
\name{tar_stan_summary_join_data}
\alias{tar_stan_summary_join_data}
\title{Join some Stan data to summary output}
\usage{
tar_stan_summary_join_data(summaries, data)
}
\arguments{
\item{summaries}{A data frame of Stan posterior summaries.}

\item{data}{List, Stan dataset.}
}
\value{
A data frame of user-friendly Stan output.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
