% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_vb_rep_summary.R
\name{tar_stan_vb_rep_summary_run}
\alias{tar_stan_vb_rep_summary_run}
\title{Run a Stan model and return only the summaries.}
\usage{
tar_stan_vb_rep_summary_run(
  stan_file,
  stan_name,
  data,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  refresh,
  init,
  save_latent_dynamics,
  output_dir,
  algorithm,
  iter,
  grad_samples,
  elbo_samples,
  eta,
  adapt_engaged,
  adapt_iter,
  tol_rel_obj,
  eval_elbo,
  output_samples,
  sig_figs,
  variables,
  summaries,
  summary_args
)
}
\arguments{
\item{stan_file}{Character, Stan model file or the lines of the file,
depending on \code{compile}.}

\item{stan_name}{Friendly suffix of the Stan model target.}

\item{data}{List of data to pass to the Stan model.}

\item{compile}{Character of length 1. If \code{"original"}, then
\code{cmdstan} will compile the source file right before running
it (or skip compilation if the binary is up to date). This
assumes the worker has access to the file. If the worker
is running on a remote computer that does not have access
to the model file, set to \code{"copy"} instead. \code{compile = "copy"}
means the pipeline will read the lines of the original Stan model file
and send them to the worker. The worker writes the lines
to a local copy and compiles the model from there, so it
no longer needs access to the original Stan model file on your
local machine. However, as a result, the Stan model re-compiles
every time the main target reruns.}

\item{seed}{\code{seed} argument to \verb{$variational()}.}

\item{refresh}{\code{refresh} argument to \verb{$variational()}.}

\item{init}{\code{init} argument to \verb{$variational()}.}

\item{save_latent_dynamics}{\code{save_latent_dynamics}
argument to \verb{$variational()}.}

\item{output_dir}{\code{output_dir} argument to \verb{$variational()}.}

\item{algorithm}{\code{algorithm} argument to \verb{$variational()}.}

\item{iter}{\code{iter} argument to \verb{$variational()}.}

\item{grad_samples}{\code{grad_samples} argument to \verb{$variational()}.}

\item{elbo_samples}{\code{elbo_samples} argument to \verb{$variational()}.}

\item{eta}{\code{eta} argument to \verb{$variational()}.}

\item{adapt_engaged}{\code{adapt_engaged} argument to \verb{$variational()}.}

\item{adapt_iter}{\code{adapt_iter} argument to \verb{$variational()}.}

\item{tol_rel_obj}{\code{tol_rel_obj} argument to \verb{$variational()}.}

\item{eval_elbo}{\code{eval_elbo} argument to \verb{$variational()}.}

\item{output_samples}{\code{output_samples} argument to \verb{$variational()}.}

\item{sig_figs}{\code{sig_figs} argument to \verb{$variational()}.}

\item{variables}{\code{variables} argument to \verb{$draws()} and \verb{$summary()}
on the \code{CmdStanVB} object.}
}
\value{
A data frame of posterior summaries.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
