% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_package.R
\docType{package}
\name{stantargets-package}
\alias{stantargets-package}
\title{targets: Targets Archetypes for Stan}
\description{
The \code{stantargets} R package is an extension to
\code{targets} and \code{cmdstanr} for Bayesian data analysis.
\code{stantargets} makes it super easy to set up useful scalable
Stan pipelines that automatically parallelize the computation
and skip expensive steps when the results are already up to date.
Minimal custom code is required, and there is no need to manually
configure branching, so usage is much easier than \code{targets} alone.
\code{stantargets} can access all of \code{cmdstanr}'s major algorithms
(MCMC, variational Bayes, and optimization) and it supports
both single-fit workflows and multi-rep simulation studies.
}
