% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_stan_gq_rep_draws.R
\name{tar_stan_gq_rep_draws_run}
\alias{tar_stan_gq_rep_draws_run}
\title{Run a Stan model and return only the summaries.}
\usage{
tar_stan_gq_rep_draws_run(
  stan_file,
  stan_name,
  stan_path,
  data,
  fitted_params,
  compile,
  quiet,
  dir,
  include_paths,
  cpp_options,
  stanc_options,
  force_recompile,
  seed,
  output_dir,
  sig_figs,
  parallel_chains,
  threads_per_chain,
  variables
)
}
\arguments{
\item{data}{List of data to pass to the Stan model.}

\item{fitted_params}{Symbol, name of a \code{CmdStanFit} object
computed in a previous target: for example, the
\verb{*_mcmc_*} target from \code{\link[=tar_stan_mcmc]{tar_stan_mcmc()}}. Must be a subclass
that \verb{$generate_quantities()} can accept as \code{fitted_params}.}

\item{compile}{Do compilation? The default is \code{TRUE}. If \code{FALSE}
compilation can be done later via the \code{\link[cmdstanr:model-method-compile]{$compile()}}
method.}

\item{quiet}{\code{quiet} argument to \verb{$compile()}.}

\item{dir}{\code{dir} argument to \verb{$compile()}.}

\item{include_paths}{\code{include_paths} argument to \verb{$compile()}.}

\item{cpp_options}{\code{cpp_options} argument to \verb{$compile()}.}

\item{stanc_options}{\code{stanc_options} argument to \verb{$compile()}.}

\item{force_recompile}{\code{force_recompile} argument to \verb{$compile()}.}

\item{seed}{\code{seed} argument to \verb{$generate_quantities()}.}

\item{output_dir}{\code{output_dir} argument to \verb{$generate_quantities()}.}

\item{sig_figs}{\code{sig_figs} argument to \verb{$generate_quantities()}.}

\item{parallel_chains}{\code{parallel_chains} argument to
\verb{$generate_quantities()}.}

\item{threads_per_chain}{\code{threads_per_chain} argument to
\verb{$generate_quantities()}.}

\item{variables}{\code{variables} argument to \verb{$draws()} and \verb{$summary()}
on the \code{CmdStanGQ} object.}
}
\value{
A data frame of posterior summaries.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
