% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{tar_stan_output}
\alias{tar_stan_output}
\title{Post-process Stan output}
\usage{
tar_stan_output(
  fit,
  output_type,
  summaries,
  summary_args,
  transform,
  variables,
  inc_warmup,
  data,
  data_copy,
  seed
)
}
\arguments{
\item{fit}{A Stan fit object.}

\item{output_type}{Type of output to create, either \code{"summaries"},
\code{"draws"}, or \code{"diagnostics"}.}

\item{summaries}{Optional list of summary functions passed to \code{...} in
\code{posterior::summarize_draws()} through \verb{$summary()}
on the \code{CmdStanFit} object.}

\item{summary_args}{Optional list of summary function arguments passed to
\code{.args} in \code{posterior::summarize_draws()} through \verb{$summary()}
on the \code{CmdStanFit} object.}

\item{transform}{Symbol or \code{NULL}, name of a function that accepts
arguments \code{data} and \code{draws} and returns a data frame. Here,
\code{data} is the JAGS data list supplied to the model, and \code{draws}
is a data frame with one column per model parameter and one row
per posterior sample. Any arguments other than \code{data} and \code{draws}
must have valid default values because \code{stantargets} will not
populate them. See the simulation-based calibration (SBC)
section of the simulation vignette for an example.}

\item{variables}{(character vector) The variables to include.}

\item{inc_warmup}{Logical, whether to include the warmup draws.}

\item{data}{List, Stan dataset.}

\item{data_copy}{Character vector of names of scalars in \code{data}.
These values will be inserted as columns in the output data frame
for each rep. To join more than just scalars, include a \code{.join_data}
element of your Stan data list with names and dimensions corresponding
to those of the model. For details, read
\url{https://docs.ropensci.org/stantargets/articles/simulation.html}.}

\item{seed}{Integer vector, random number generator seed used to run Stan.}
}
\value{
A data frame of user-friendly Stan output.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
